接口自动化框架/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
iniconfig==1.1.1
jsonpath==0.82
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==7.0.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
tomli==1.2.3
urllib3==1.26.8
